+++ gen/com/yanncebron/m68kplugin/lang/M68kParser.java
@@ -79,17 +79,17 @@ public class M68kParser implements PsiParser, LightPsiParser {
       EXG_INSTRUCTION, EXT_INSTRUCTION, LEA_INSTRUCTION, LINK_INSTRUCTION,
       LSL_INSTRUCTION, LSR_INSTRUCTION, MACRO_CALL_DIRECTIVE, MOVEA_INSTRUCTION,
       MOVEC_INSTRUCTION, MOVEM_INSTRUCTION, MOVEP_INSTRUCTION, MOVEQ_INSTRUCTION,
-      MOVE_INSTRUCTION, MULS_INSTRUCTION, MULU_INSTRUCTION, NBCD_INSTRUCTION,
-      NEGX_INSTRUCTION, NEG_INSTRUCTION, NOT_INSTRUCTION, ORI_INSTRUCTION,
-      OR_INSTRUCTION, PEA_INSTRUCTION, ROL_INSTRUCTION, ROR_INSTRUCTION,
-      ROXL_INSTRUCTION, ROXR_INSTRUCTION, RS_DIRECTIVE, SBCD_INSTRUCTION,
-      SCC_INSTRUCTION, SCS_INSTRUCTION, SEQ_INSTRUCTION, SF_INSTRUCTION,
-      SGE_INSTRUCTION, SGT_INSTRUCTION, SHI_INSTRUCTION, SHS_INSTRUCTION,
-      SLE_INSTRUCTION, SLO_INSTRUCTION, SLS_INSTRUCTION, SLT_INSTRUCTION,
-      SMI_INSTRUCTION, SNE_INSTRUCTION, SPL_INSTRUCTION, ST_INSTRUCTION,
-      SUBA_INSTRUCTION, SUBI_INSTRUCTION, SUBQ_INSTRUCTION, SUBX_INSTRUCTION,
-      SUB_INSTRUCTION, SVC_INSTRUCTION, SVS_INSTRUCTION, SWAP_INSTRUCTION,
-      TAS_INSTRUCTION, TST_INSTRUCTION),
+      MOVES_INSTRUCTION, MOVE_INSTRUCTION, MULS_INSTRUCTION, MULU_INSTRUCTION,
+      NBCD_INSTRUCTION, NEGX_INSTRUCTION, NEG_INSTRUCTION, NOT_INSTRUCTION,
+      ORI_INSTRUCTION, OR_INSTRUCTION, PEA_INSTRUCTION, ROL_INSTRUCTION,
+      ROR_INSTRUCTION, ROXL_INSTRUCTION, ROXR_INSTRUCTION, RS_DIRECTIVE,
+      SBCD_INSTRUCTION, SCC_INSTRUCTION, SCS_INSTRUCTION, SEQ_INSTRUCTION,
+      SF_INSTRUCTION, SGE_INSTRUCTION, SGT_INSTRUCTION, SHI_INSTRUCTION,
+      SHS_INSTRUCTION, SLE_INSTRUCTION, SLO_INSTRUCTION, SLS_INSTRUCTION,
+      SLT_INSTRUCTION, SMI_INSTRUCTION, SNE_INSTRUCTION, SPL_INSTRUCTION,
+      ST_INSTRUCTION, SUBA_INSTRUCTION, SUBI_INSTRUCTION, SUBQ_INSTRUCTION,
+      SUBX_INSTRUCTION, SUB_INSTRUCTION, SVC_INSTRUCTION, SVS_INSTRUCTION,
+      SWAP_INSTRUCTION, TAS_INSTRUCTION, TST_INSTRUCTION),
   };
 
   /* ********************************************************** */
@@ -2725,7 +2725,8 @@ public class M68kParser implements PsiParser, LightPsiParser {
   //                               moveq_instruction |
   //                               movem_instruction |
   //                               movep_instruction |
-  //                               movec_instruction
+  //                               movec_instruction |
+  //                               moves_instruction
   static boolean move_instructions(PsiBuilder b, int l) {
     if (!recursion_guard_(b, l, "move_instructions")) return false;
     boolean r;
@@ -2735,6 +2736,7 @@ public class M68kParser implements PsiParser, LightPsiParser {
     if (!r) r = movem_instruction(b, l + 1);
     if (!r) r = movep_instruction(b, l + 1);
     if (!r) r = movec_instruction(b, l + 1);
+    if (!r) r = moves_instruction(b, l + 1);
     return r;
   }
 
@@ -3127,6 +3129,68 @@ public class M68kParser implements PsiParser, LightPsiParser {
     if (!recursion_guard_(b, l, "moveq_instruction_1")) return false;
     data_size_long(b, l + 1);
     return true;
+  }
+
+  /* ********************************************************** */
+  // MOVES data_size_all?
+  //                       (
+  //                         moves_tail_from |
+  //                         moves_tail_to
+  //                       )
+  public static boolean moves_instruction(PsiBuilder b, int l) {
+    if (!recursion_guard_(b, l, "moves_instruction")) return false;
+    if (!nextTokenIs(b, "<instruction>", MOVES)) return false;
+    boolean r, p;
+    Marker m = enter_section_(b, l, _NONE_, MOVES_INSTRUCTION, "<instruction>");
+    r = consumeToken(b, MOVES);
+    p = r; // pin = 1
+    r = r && report_error_(b, moves_instruction_1(b, l + 1));
+    r = p && moves_instruction_2(b, l + 1) && r;
+    exit_section_(b, l, m, r, p, null);
+    return r || p;
+  }
+
+  // data_size_all?
+  private static boolean moves_instruction_1(PsiBuilder b, int l) {
+    if (!recursion_guard_(b, l, "moves_instruction_1")) return false;
+    data_size_all(b, l + 1);
+    return true;
+  }
+
+  // moves_tail_from |
+  //                         moves_tail_to
+  private static boolean moves_instruction_2(PsiBuilder b, int l) {
+    if (!recursion_guard_(b, l, "moves_instruction_2")) return false;
+    boolean r;
+    r = moves_tail_from(b, l + 1);
+    if (!r) r = moves_tail_to(b, l + 1);
+    return r;
+  }
+
+  /* ********************************************************** */
+  // adm_group_all_except_ard COMMA adm_rrd
+  static boolean moves_tail_from(PsiBuilder b, int l) {
+    if (!recursion_guard_(b, l, "moves_tail_from")) return false;
+    boolean r;
+    Marker m = enter_section_(b);
+    r = adm_group_all_except_ard(b, l + 1);
+    r = r && consumeToken(b, COMMA);
+    r = r && adm_rrd(b, l + 1);
+    exit_section_(b, m, null, r);
+    return r;
+  }
+
+  /* ********************************************************** */
+  // adm_rrd COMMA adm_group_all_except_ard
+  static boolean moves_tail_to(PsiBuilder b, int l) {
+    if (!recursion_guard_(b, l, "moves_tail_to")) return false;
+    boolean r;
+    Marker m = enter_section_(b);
+    r = adm_rrd(b, l + 1);
+    r = r && consumeToken(b, COMMA);
+    r = r && adm_group_all_except_ard(b, l + 1);
+    exit_section_(b, m, null, r);
+    return r;
   }
 
   /* ********************************************************** */