+++ src/com/yanncebron/m68kplugin/lang/psi/M68kTokenGroups.java
@@ -127,6 +127,7 @@ public final class M68kTokenGroups {
     MOVEM,
     MOVEP,
     MOVEQ,
+    MOVES,
     MULS,
     MULU,
     NBCD,