+++ tests/com/yanncebron/m68kplugin/parser/MnemonicGeneratedParserDataTest.java
@@ -74,7 +74,7 @@ public class MnemonicGeneratedParserDataTest extends M68kParsingTestCase {
 
     dump("* Instructions count: " + M68kTokenGroups.INSTRUCTIONS.getTypes().length);
 
-    assertEquals(5346, total);
+    assertEquals(5506, total);
     assertEmpty(failedVariants);
     assertEquals(781, deprecated);
   }